gogensig:move full test to folder test
