+++ chore/gogensig/convert/_testdata/avoidkeyword/conf/llcppg.cfg
@@ -1,0 +1,5 @@
+{
+  "name": "avoidkeyword",
+  "include": ["temp.h"],
+  "cplusplus":false
+}
