+++ chore/gogensig/convert/_testdata/avoidkeyword/conf/llcppg.symb.json
@@ -1,0 +1,7 @@
+[
+  {
+    "mangle": "lua_sethook",
+    "c++": "lua_sethook",
+    "go": "Sethook"
+  }
+]
