+++ chore/gogensig/convert/_testdata/avoidkeyword/gogensig.expect
@@ -1,0 +1,19 @@
+===== temp.go =====
+package avoidkeyword
+
+import (
+	"github.com/goplus/llgo/c"
+	_ "unsafe"
+)
+
+type LuaState struct {
+	Unused [8]uint8
+}
+// llgo:type C
+type LuaHook func(*LuaState)
+//go:linkname Sethook C.lua_sethook
+func Sethook(L *LuaState, func_ LuaHook, mask c.Int, count c.Int)
+
+===== llcppg.pub =====
+lua_Hook LuaHook
+lua_State LuaState
