+++ chore/gogensig/convert/_testdata/enum/conf/llcppg.cfg
@@ -1,0 +1,5 @@
+{
+  "name": "enum",
+  "include": ["temp.h"],
+  "cplusplus":false
+}
