+++ chore/gogensig/convert/_testdata/enum/hfile/temp.h
@@ -1,0 +1,8 @@
+enum { enum1, enum2 };
+enum spectrum { red, orange, yello, green, blue, violet };
+
+enum kids { nippy, slats, skippy, nina, liz };
+
+enum levels { low = 100, medium = 500, high = 2000 };
+
+enum feline { cat, lynx = 10, puma, tiger };
