+++ chore/gogensig/convert/_testdata/forwarddecl/conf/llcppg.cfg
@@ -1,0 +1,6 @@
+{
+  "name": "forwarddecl",
+  "include": ["temp.h"],
+  "trimPrefixes": ["sqlite3_"],
+  "cplusplus":false
+}
