+++ chore/gogensig/convert/_testdata/keepcomment/conf/llcppg.cfg
@@ -1,0 +1,5 @@
+{
+  "name": "comment",
+  "include": ["temp.h"],
+  "cplusplus":false
+}
