+++ chore/gogensig/convert/_testdata/keepcomment/conf/llcppg.symb.json
@@ -1,0 +1,7 @@
+[
+  {
+    "mangle": "ExecuteFoo",
+    "c++": "ExecuteFoo",
+    "go": "CustomExecuteFoo"
+  }
+]
