+++ chore/gogensig/convert/_testdata/keepcomment/hfile/temp.h
@@ -1,0 +1,9 @@
+/// Foo comment
+struct Foo {
+    int a;
+    double b;
+    bool c;
+};
+
+/// ExecuteFoo comment
+int ExecuteFoo(int a, Foo b);
