+++ chore/gogensig/convert/_testdata/pubfile/conf/llcppg.cfg
@@ -1,0 +1,5 @@
+{
+  "name": "pubfile",
+  "include": ["temp.h"],
+  "cplusplus":false
+}
