+++ chore/gogensig/convert/_testdata/pubfile/conf/llcppg.pub
@@ -1,0 +1,1 @@
+data CustomData
