+++ chore/gogensig/convert/_testdata/pubfile/conf/llcppg.symb.json
@@ -1,0 +1,7 @@
+[
+  {
+    "mangle": "func",
+    "c++": "func",
+    "go": "Func"
+  }
+]
