+++ chore/gogensig/convert/_testdata/pubfile/hfile/temp.h
@@ -1,0 +1,17 @@
+struct point {
+    int x;
+    int y;
+};
+struct Capital {
+    int x;
+    int y;
+};
+union data {
+    float f;
+    char str[20];
+};
+typedef unsigned int uint_t;
+enum color {
+    RED = 0,
+};
+void func(int a, int b);
