+++ chore/gogensig/convert/_testdata/selfref/conf/llcppg.cfg
@@ -1,0 +1,5 @@
+{
+  "name": "selfref",
+  "include": ["temp.h"],
+  "cplusplus":false
+}
