+++ chore/gogensig/convert/_testdata/selfref/hfile/temp.h
@@ -1,0 +1,10 @@
+typedef struct cJSON {
+    struct cJSON *next;
+    struct cJSON *prev;
+    struct cJSON *child;
+    int type;
+    char *valuestring;
+    int valueint;
+    double valuedouble;
+    char *string;
+} cJSON;
