+++ chore/gogensig/convert/_testdata/stdtype/conf/llcppg.cfg
@@ -1,0 +1,5 @@
+{
+  "name": "stdtype",
+  "include": ["temp.h"],
+  "cplusplus":false
+}
