+++ chore/gogensig/convert/_testdata/stdtype/conf/llcppg.symb.json
@@ -1,0 +1,7 @@
+[
+  {
+    "mangle": "testStdType",
+    "c++": "testStdType",
+    "go": "TestStdType"
+  }
+]
