+++ chore/gogensig/convert/_testdata/stdtype/hfile/temp.h
@@ -1,0 +1,3 @@
+#include <stddef.h>
+
+void testStdType(size_t a, ptrdiff_t b);
