+++ chore/gogensig/convert/_testdata/union/conf/llcppg.cfg
@@ -1,0 +1,5 @@
+{
+  "name": "enum",
+  "include": ["temp.h"],
+  "cplusplus":false
+}
