+++ chore/gogensig/convert/_testdata/union/hfile/temp.h
@@ -1,0 +1,5 @@
+typedef union __u {
+    int a;
+    long b;
+    float c;
+} u;
